+++ tests/spec/non_conformant/mixin/content/arguments/error/syntax.rs
@@ -28,7 +28,7 @@ mod arglist {
         );
     }
     #[test]
-    #[ignore] // wrong error
+    #[ignore] // missing error
     fn missing() {
         assert_eq!(
             runner().err(